requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyMySQL==1.1.0
python-dotenv==1.0.1
bcrypt
PyJWT==2.8.0
